♻️ Refactor URL parameter extraction into a dedicated utility module
