+++ frontend/packages/erd-core/src/components/ERDRenderer/ERDContent/useInitialAutoLayout.ts
@@ -1,49 +1,18 @@
-import type { QueryParam } from '@/schemas/queryParam'
-import { type ShowMode, showModeSchema } from '@/schemas/showMode'
 import {
   addHiddenNodeIds,
   updateActiveTableName,
   updateShowMode,
 } from '@/stores'
-import { decompressFromEncodedURIComponent } from '@/utils'
+import {
+  getActiveTableNameFromUrl,
+  getHiddenNodeIdsFromUrl,
+  getShowModeFromUrl,
+} from '@/utils'
 import { type Node, useReactFlow } from '@xyflow/react'
 import { useEffect, useMemo } from 'react'
-import * as v from 'valibot'
 import { useERDContentContext } from './ERDContentContext'
 import { highlightNodesAndEdges } from './highlightNodesAndEdges'
 import { useAutoLayout } from './useAutoLayout'
-
-const getActiveTableNameFromUrl = (): string | undefined => {
-  const urlParams = new URLSearchParams(window.location.search)
-  const activeQueryParam: QueryParam = 'active'
-  const tableName = urlParams.get(activeQueryParam)
-
-  return tableName || undefined
-}
-
-export const getHiddenNodeIdsFromUrl = async (): Promise<string[]> => {
-  const urlParams = new URLSearchParams(window.location.search)
-  const hiddenQueryParam: QueryParam = 'hidden'
-  const compressed = urlParams.get(hiddenQueryParam)
-  const hiddenNodeIds = compressed
-    ? await decompressFromEncodedURIComponent(compressed).catch(() => undefined)
-    : undefined
-
-  return hiddenNodeIds ? hiddenNodeIds.split(',') : []
-}
-
-const getShowModeFromUrl = (): ShowMode => {
-  const urlParams = new URLSearchParams(window.location.search)
-  const showModeQueryParam: QueryParam = 'showMode'
-  const showMode = urlParams.get(showModeQueryParam)
-
-  const result = v.safeParse(showModeSchema, showMode)
-  if (result.success && result.output) {
-    return result.output
-  }
-
-  return 'TABLE_NAME'
-}
 
 export const useInitialAutoLayout = (
   nodes: Node[],
